Ошибки

+++ fastapi_app/src/use_cases/user/delete_user.py
@@ -1,4 +1,5 @@
 from src.repositories.user_repository import UserRepository
+from src.exceptions import NotFoundError
 
 
 class DeleteUserUseCase:
@@ -8,4 +9,8 @@ class DeleteUserUseCase:
         self.repository = repository
     
     def execute(self, user_id: int) -> bool:
+        existing_user = self.repository.get_by_id(user_id)
+        if not existing_user:
+            raise NotFoundError("User", user_id)
+        
         return self.repository.delete(user_id)
